+++ src/mocks/ERC721Mock.sol
@@ -10,15 +10,20 @@ contract ERC721Mock is ERC721 {
 
     string public baseTokenURI;
 
-    function mint(address to, uint tokenId) public {
+    function mint(address to, uint256 tokenId) public {
         _safeMint(to, tokenId, "");
     }
 
-    function transfer(address to, uint tokenId) public {
+    function transfer(address to, uint256 tokenId) public {
         _safeTransfer(msg.sender, to, tokenId, "");
     }
 
-    function isApprovedOrOwner(address spender, uint tokenId) public view virtual returns (bool) {
+    function isApprovedOrOwner(address spender, uint256 tokenId)
+        public
+        view
+        virtual
+        returns (bool)
+    {
         return _isApprovedOrOwner(spender, tokenId);
     }
 }
