+++ src/token/onft/ONFT721Core.sol
@@ -9,21 +9,59 @@ import "@openzeppelin/contracts/utils/introspection/ERC165.sol";
 abstract contract ONFT721Core is NonblockingLzApp, ERC165, IONFT721Core {
     constructor(address _lzEndpoint) NonblockingLzApp(_lzEndpoint) {}
 
-    function supportsInterface(bytes4 interfaceId) public view virtual override(ERC165, IERC165) returns (bool) {
-        return interfaceId == type(IONFT721Core).interfaceId || super.supportsInterface(interfaceId);
+    function supportsInterface(bytes4 interfaceId)
+        public
+        view
+        virtual
+        override(ERC165, IERC165)
+        returns (bool)
+    {
+        return
+            interfaceId == type(IONFT721Core).interfaceId || super.supportsInterface(interfaceId);
     }
 
-    function estimateSendFee(uint16 _dstChainId, bytes memory _toAddress, uint _tokenId, bool _useZro, bytes memory _adapterParams) public view virtual override returns (uint nativeFee, uint zroFee) {
+    function estimateSendFee(
+        uint16 _dstChainId,
+        bytes memory _toAddress,
+        uint256 _tokenId,
+        bool _useZro,
+        bytes memory _adapterParams
+    ) public view virtual override returns (uint256 nativeFee, uint256 zroFee) {
         // mock the payload for send()
         bytes memory payload = abi.encode(_toAddress, _tokenId);
-        return lzEndpoint.estimateFees(_dstChainId, address(this), payload, _useZro, _adapterParams);
+        return
+            lzEndpoint.estimateFees(_dstChainId, address(this), payload, _useZro, _adapterParams);
     }
 
-    function sendFrom(address _from, uint16 _dstChainId, bytes memory _toAddress, uint _tokenId, address payable _refundAddress, address _zroPaymentAddress, bytes memory _adapterParams) public payable virtual override {
-        _send(_from, _dstChainId, _toAddress, _tokenId, _refundAddress, _zroPaymentAddress, _adapterParams);
+    function sendFrom(
+        address _from,
+        uint16 _dstChainId,
+        bytes memory _toAddress,
+        uint256 _tokenId,
+        address payable _refundAddress,
+        address _zroPaymentAddress,
+        bytes memory _adapterParams
+    ) public payable virtual override {
+        _send(
+            _from,
+            _dstChainId,
+            _toAddress,
+            _tokenId,
+            _refundAddress,
+            _zroPaymentAddress,
+            _adapterParams
+        );
     }
 
-    function _send(address _from, uint16 _dstChainId, bytes memory _toAddress, uint _tokenId, address payable _refundAddress, address _zroPaymentAddress, bytes memory _adapterParams) internal virtual {
+    function _send(
+        address _from,
+        uint16 _dstChainId,
+        bytes memory _toAddress,
+        uint256 _tokenId,
+        address payable _refundAddress,
+        address _zroPaymentAddress,
+        bytes memory _adapterParams
+    ) internal virtual {
         _debitFrom(_from, _dstChainId, _toAddress, _tokenId);
 
         bytes memory payload = abi.encode(_toAddress, _tokenId);
@@ -33,9 +71,14 @@ abstract contract ONFT721Core is NonblockingLzApp, ERC165, IONFT721Core {
         emit SendToChain(_from, _dstChainId, _toAddress, _tokenId, nonce);
     }
 
-    function _nonblockingLzReceive(uint16 _srcChainId, bytes memory _srcAddress, uint64 _nonce, bytes memory _payload) internal virtual override {
+    function _nonblockingLzReceive(
+        uint16 _srcChainId,
+        bytes memory _srcAddress,
+        uint64 _nonce,
+        bytes memory _payload
+    ) internal virtual override {
         // decode and load the toAddress
-        (bytes memory toAddressBytes, uint tokenId) = abi.decode(_payload, (bytes, uint));
+        (bytes memory toAddressBytes, uint256 tokenId) = abi.decode(_payload, (bytes, uint256));
         address toAddress;
         assembly {
             toAddress := mload(add(toAddressBytes, 20))
@@ -46,7 +89,16 @@ abstract contract ONFT721Core is NonblockingLzApp, ERC165, IONFT721Core {
         emit ReceiveFromChain(_srcChainId, _srcAddress, toAddress, tokenId, _nonce);
     }
 
-    function _debitFrom(address _from, uint16 _dstChainId, bytes memory _toAddress, uint _tokenId) internal virtual;
+    function _debitFrom(
+        address _from,
+        uint16 _dstChainId,
+        bytes memory _toAddress,
+        uint256 _tokenId
+    ) internal virtual;
 
-    function _creditTo(uint16 _srcChainId, address _toAddress, uint _tokenId) internal virtual;
+    function _creditTo(
+        uint16 _srcChainId,
+        address _toAddress,
+        uint256 _tokenId
+    ) internal virtual;
 }
